After the council of debugging.

+++ src/context/BookingContext.js
@@ -115,20 +115,17 @@ const BookingProvider = (props) => {
 
   //Set Local storage to remember last booking
   useEffect(() => {
-    if (confirmationDetails) {
-      localStorage.setItem("lastBooking", JSON.stringify([...confirmationDetails, currentUser]))
-    }
-
-  }, [confirmationDetails])
+    if (confirmationDetails) localStorage.setItem("lastBooking", JSON.stringify([...confirmationDetails, currentUser]));
+  }, [confirmationDetails, currentUser]);
 
   //Get from Local storage on hard reload and when user updates.
   useEffect(() => {
     let storageItem = JSON.parse(localStorage.getItem("lastBooking"));
 
-    if (currentUser && storageItem && currentUser._id == storageItem[2]._id) {
-      setConfirmationDetails([storageItem[0], storageItem[1]])
+    if (currentUser && currentUser.bookings.length != 0 && !confirmationDetails && storageItem && currentUser._id == storageItem[2]._id) {
+      setConfirmationDetails([storageItem[0], storageItem[1]]);
     }
-  }, [currentUser])
+  }, [currentUser]);
 
   const values = {
     bookingsId,
